aindump: guess filenames if no FNAM section
If the filenames section isn't present, try to guess the filenames from
the functions defined in the file. This works pretty well for most of
the core engine files but gives pretty crappy results for scenario
files.

This feature is primarily intended to be used when dumping the code
section to multiple files (not yet implemented).