parts: fix render order of child parts
The parent's z-value determines the render order of its children.

Alicesoft's implementation of this is a bit weird. The render order only
changes when the parent's visibility is toggled (via Parts_SetShow), NOT
when the parent is set on a child or the parent's z-value changes.

I have chosen to implement this so that the render order is always based
on the current parent/child z-values.

This fixes some bugs in Rance 01.