Windows: Fix read_save_file
This fixes an issue where saved files could not be read on Windows.

On Windows if a file is opened in text mode, ftell() at EOF returns
the file size in bytes, but fread() tries to read the specified number
of characters (after removing carriage returns), resulting in an EOF
error.

To avoid this problem, open save files in binary mode.